move ender link related classes to own package
